Move CurrencyInputValidator
