+++ core/interactors/src/commonMain/kotlin/org.pointyware.xyz.core.interactors/business/CurrencyInputValidator.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.drive.interactors
+package org.pointyware.xyz.core.interactors.business
 
 import org.pointyware.xyz.core.entities.business.Currency
